refactor: 基底クラスを作成し、共通の処理は基底クラスで行うようにする
