+++ lib/tenji_maker/middle.rb
@@ -1,27 +1,5 @@
 class TenjiMaker
-  class Middle
-    attr_reader :text
-
-    def initialize(text)
-      @text = text.split(" ")
-    end
-
-    def position
-      text.map do |char|
-        next for_N if char == "N"
-
-        if none?(char)
-          :none
-        elsif right?(char)
-          :right
-        elsif left?(char)
-          :left
-        else
-          :all
-        end
-      end
-    end
-
+  class Middle < Base
     private
 
     def for_N
